chore: Move patchRequest to shared
