+++ .changeset/cyan-needles-listen.md
@@ -1,5 +1,6 @@
 ---
 "@clerk/astro": patch
+"@clerk/shared": patch
 ---
 
 Fixed a bug where the `clerkMiddleware()` helper would consume the body of the request.
